Add WriteOff works
Updated possibility of adding and updating writeOff, not finished

+++ Tailor/src/xyzy/model/Product.java
@@ -17,7 +17,7 @@ public class Product {
     public void setID(int ID) {
         this.ID = ID;
     }
-    public long getID() {
+    public int getID() {
         return ID;
     }
